Use the max7456 VSYNC signal to initial writing to the display to avoid display artifacts.
This also relocates quite a lot of low level code into the video driver
from the high level osd_max7456.c
